Improve type hierarchy.
