+++ common/src/main/java/io/github/reserveword/imblocker/common/gui/MinecraftAbstractTextInputWidget.java
@@ -1,0 +1,30 @@
+package io.github.reserveword.imblocker.common.gui;
+
+import io.github.reserveword.imblocker.common.IMManager;
+
+public interface MinecraftAbstractTextInputWidget<CI extends CursorInfo> extends MinecraftFocusableWidget {
+	
+	@Override
+	default void deliverFocus() {
+		updateCursorInfo();
+		MinecraftFocusableWidget.super.deliverFocus();
+	}
+	
+	default void imblocker$onBoundsChanged() {
+		if(isTrulyFocused()) {
+			IMManager.updateCompositionWindowPos();
+		}
+	}
+	
+	default void imblocker$onCursorChanged() {
+		if(isTrulyFocused() && updateCursorInfo()) {
+			IMManager.updateCompositionWindowPos();
+		}
+	}
+	
+	default boolean updateCursorInfo() {
+		return true;
+	}
+
+	CI getCursorInfo();
+}
