+++ common/src/main/java/io/github/reserveword/imblocker/common/gui/MinecraftMultilineEditBoxWidget.java
@@ -1,22 +1,9 @@
 package io.github.reserveword.imblocker.common.gui;
 
-import io.github.reserveword.imblocker.common.IMManager;
 import io.github.reserveword.imblocker.common.StringUtil;
 import io.github.reserveword.imblocker.common.accessor.MinecraftClientAccessor;
 
-public interface MinecraftMultilineEditBoxWidget extends MinecraftFocusableWidget {
-	
-	@Override
-	default void deliverFocus() {
-		updateCursorInfo();
-		MinecraftFocusableWidget.super.deliverFocus();
-	}
-	
-	default void imblocker$onBoundsChanged() {
-		if(isTrulyFocused()) {
-			IMManager.updateCompositionWindowPos();
-		}
-	}
+public interface MinecraftMultilineEditBoxWidget extends MinecraftAbstractTextInputWidget<MultilineCursorInfo> {
 	
 	default Point getCaretPos() {
 		MultilineCursorInfo cursorInfo = getCursorInfo();
@@ -25,16 +12,4 @@ public interface MinecraftMultilineEditBoxWidget extends MinecraftFocusableWidge
 		int caretY = (int) (4 + cursorInfo.cursorLineIndex * 9 - cursorInfo.scrollY);
 		return new Point(getGuiScale(), caretX, caretY);
 	}
-	
-	default void imblocker$onCursorChanged() {
-		if(updateCursorInfo() && isTrulyFocused()) {
-			IMManager.updateCompositionWindowPos();
-		}
-	}
-	
-	default boolean updateCursorInfo() {
-		return true;
-	}
-	
-	MultilineCursorInfo getCursorInfo();
 }
